# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "brandify",
        "description": "This is the API documentation for My API.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/comments/create": {
            "get": {
                "summary": "Show the form for creating a new comment",
                "operationId": "0733ca7870a2fc4c1610d473c6dca069",
                "responses": {
                    "200": {
                        "description": "Form to create a new comment"
                    }
                }
            }
        },
        "/api/comments": {
            "post": {
                "summary": "Store a newly created comment in storage",
                "operationId": "300619a0a4003920aafe960e45688102",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Comment"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Comment created successfully"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/comments/{id}/edit": {
            "get": {
                "summary": "Show the form for editing the specified comment",
                "operationId": "2bb05d87e8dd21d0020209b7893a721f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Form to edit a comment"
                    },
                    "404": {
                        "description": "Comment not found"
                    }
                }
            }
        },
        "/api/comments/{id}": {
            "put": {
                "summary": "Update the specified comment in storage",
                "operationId": "39e62963cbf897a1b2168683878e0a46",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Comment"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Comment updated successfully"
                    },
                    "404": {
                        "description": "Comment not found"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            },
            "delete": {
                "summary": "Remove the specified comment from storage",
                "operationId": "721e5dc0c32ef81276203df965767606",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Comment deleted successfully"
                    },
                    "404": {
                        "description": "Comment not found"
                    }
                }
            }
        },
        "/api/posts": {
            "get": {
                "summary": "List all posts",
                "operationId": "e9eaf7df702a94f4443677bd1a36cfe6",
                "responses": {
                    "200": {
                        "description": "A list of posts"
                    }
                }
            }
        },
        "/api/posts/create": {
            "get": {
                "summary": "Show the form for creating a new post",
                "operationId": "f5b30e46f644a81e3fe49f47ef075125",
                "responses": {
                    "200": {
                        "description": "Form to create a new post"
                    }
                }
            }
        },
        "/posts": {
            "post": {
                "summary": "Create a new post",
                "operationId": "c2e0b0e26215e6d34ece33ebae25f7df",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/posts/{id}": {
            "get": {
                "summary": "Display the specified post",
                "operationId": "b0f9d92879218f68a376b6edc6467c67",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Display a single post"
                    },
                    "404": {
                        "description": "Post not found"
                    }
                }
            },
            "delete": {
                "summary": "Remove the specified post from storage",
                "operationId": "fad1726bd24cd459fc1525b67985df39",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post deleted successfully"
                    },
                    "404": {
                        "description": "Post not found"
                    }
                }
            }
        },
        "/posts/{id}/edit": {
            "get": {
                "summary": "Edit an existing post",
                "operationId": "21af8f810c09713b3cf9987a1624e4fd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the post to edit",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful"
                    }
                }
            }
        },
        "/posts/{id}": {
            "put": {
                "summary": "Update an existing post",
                "operationId": "06c81121178208730e11d2f6b655e78d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the post to update",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Comment": {
                "required": [
                    "content",
                    "post_id",
                    "user_id"
                ],
                "properties": {
                    "content": {
                        "description": "The content of the comment",
                        "type": "string"
                    },
                    "post_id": {
                        "description": "The ID of the post the comment belongs to",
                        "type": "integer"
                    },
                    "user_id": {
                        "description": "The ID of the user who created the comment",
                        "type": "integer"
                    },
                    "parent_id": {
                        "description": "The ID of the parent comment if this is a reply",
                        "type": "integer",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        }
    }
}